test: merge test files and add bun test coverage script with 98% threshold
