+++ bunfig.toml
@@ -1,0 +1,3 @@
+[test]
+coverageThreshold = 0.98
+coverageSkipTestFiles = true
