+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "zephyr",
-  "version": "1.0.8",
+  "version": "1.0.9",
   "license": "AGPL-3.0",
   "homepage": "https://zephyyrr.in",
   "private": true,
@@ -19,6 +19,7 @@
     "db:up": "bun run --filter=@zephyr/db generate && bun run --filter=@zephyr/db push",
     "build": "turbo check-types && turbo build",
     "test": "bun test",
+    "test:coverage": "bun test --coverage",
     "dep:check": "bunx sherif -f",
     "search:init": "cd packages/db && bunx tsx scripts/init-meilisearch.ts",
     "docker:dev": "docker compose -f docker/docker-compose.dev.yml --profile init --profile studio up -d",
